% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_single_instance_prices.R
\name{analyze_single_instance_data}
\alias{analyze_single_instance_data}
\title{Analyze Single Instance Data}
\usage{
analyze_single_instance_data(
  instance_df,
  output_pdf_path = "single_instance_data.pdf"
)
}
\arguments{
\item{instance_df}{A dataframe of locational marginal prices with LMP names}

\item{output_pdf_path}{The path to save the pdf file. There is a default path of "single_instance_data.pdf" but the user can edit this path to save the file in a different location.}
}
\value{
A pdf file with the summary statistics and violin plot
}
\description{
This function calls the functions \code{summarize_lmps} and \code{price_violin} with an inputting dataframe of an instance of data (one date and 15 minute interval) and combines the results into a pdf to export. First, the user has to call the \code{pulldata_instance} function to get the data for the instance they want to analyze.
}
