% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_two_node_prices.R
\name{analyze_two_node_data}
\alias{analyze_two_node_data}
\title{Analyze Two Node Data}
\usage{
analyze_two_node_data(
  df1 = read.csv("inst/extdata/San_Francisco_node_2023.csv"),
  df2 = read.csv("inst/extdata/Los_Angeles_node_2023.csv"),
  output_pdf_path = "two_node_data.pdf"
)
}
\arguments{
\item{df1}{A dataframe of locational marginal prices for a year for one location. The default is San Francisco for 2023. Users can change this dataframe to analyze a different location or year by pulling from the \code{pulldata_node} function}

\item{df2}{A dataframe of locational marginal prices for a year for another location. The defualt is Los Angeles for 2023. Users can change this dataframe to analyze a different location or year by pulling from the \code{pulldata_node} function}

\item{output_pdf_path}{The path to save the pdf file. There is a default path of "single_instance_data.pdf" but the user can edit this path to save the file in a different location.}
}
\value{
A pdf file with the plots
}
\description{
This function calls the functions \code{compare_lmps_visual}, \code{compare_lmps_visual_seasonal}, \code{compare_congestion_visual}, \code{compare_congestion_visual_seasonal}, \code{compare_losses_visual}, and \code{compare_losses_visual_seasonal} with two inputting dataframes of two nodes of data (two locations over a year) and combines the results into a pdf to export. First, the user has to call the \code{pulldata_node} function twice to get the data for the two nodes they want to analyze, or they can use the default.
}
