% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caiso_api_wrapper.R
\name{build_api_request}
\alias{build_api_request}
\title{Construct API Request}
\usage{
build_api_request(client, endpoint, params = list())
}
\arguments{
\item{client}{CAISO client object}

\item{endpoint}{API endpoint}

\item{params}{Query parameters}
}
\value{
httr2 request object
}
\description{
Build a robust API request with error handling
}
