% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA_map.R
\name{CA_map}
\alias{CA_map}
\title{create base map of NE}
\usage{
CA_map(instance = (read.csv("data/instance_jan1_2020.csv")))
}
\value{
a map of the Northeastern states of the United States
}
\description{
function creates a map of the Northeastern states of the United States
}
\examples{
CA_map()

}
