% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caiso_api_wrapper.R, R/isone_api_wrapper.R
\name{CAISOClient}
\alias{CAISOClient}
\title{ISONE API Wrapper Constructor}
\usage{
CAISOClient(
  host = "https://api.gridstatus.io/v1",
  api_key = NULL,
  max_retries = 3
)

CAISOClient(
  host = "https://api.gridstatus.io/v1",
  api_key = NULL,
  max_retries = 3
)
}
\arguments{
\item{host}{The base URL of the Grid Status API. Default is "https://api.gridstatus.io/v1".}

\item{api_key}{The API key to use for authenticating requests to the API. If not provided, the function will
attempt to read the API key from the \code{GRIDSTATUS_API_KEY} environment variable.}

\item{max_retries}{The maximum number of times to retry a request if it fails. Default is 3.}
}
\value{
An CAISOClient object.

An CAISOClient object.
}
\description{
Construct an CAISOClient object for interacting with the Grid Status API.

Construct an CAISOClient object for interacting with the Grid Status API.
}
\details{
The CAISOClient object is used to interact with the Grid Status API. In order to fetch LMP data
from the API, need to generate a free api key from the Grid Status website. You can do so by
signing up for an account at https://gridstatus.io/signup. Once you have an API key, you can
pass it to the CAISOClient constructor or set the environment variable \code{GRIDSTATUS_API_KEY} to
create a client object that can be used to fetch data.

The CAISOClient object is used to interact with the Grid Status API. In order to fetch LMP data
from the API, need to generate a free api key from the Grid Status website. You can do so by
signing up for an account at https://gridstatus.io/signup. Once you have an API key, you can
pass it to the CAISOClient constructor or set the environment variable \code{GRIDSTATUS_API_KEY} to
create a client object that can be used to fetch data.
}
