% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAISOWrapR-package.R
\docType{package}
\name{CAISOWrapR-package}
\alias{CAISOWrapR}
\alias{CAISOWrapR-package}
\title{CAISOWrapR: CAISO API Wrapper using Grid Status API}
\description{
This package provides a wrapper for the CAISO locational marginal price (LMP) datasets using the Grid Status API. The package allows users to access node-level LMP data concerning the California eletrical grid and integrate it with data concerning California's transmission lines to identity areas of high congestion. Additionally, the package provides tools to visualize and analyze the LMP data to produce insightful and effective results concerning California eletrical grid congestion.
}
\author{
\strong{Maintainer}: Mack Smith \email{macks26@stanford.edu}

Other contributors:
\itemize{
  \item First Last
  \item First Last
}

}
\keyword{internal}
