% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_two_node_prices.R
\name{compare_congestion_visual}
\alias{compare_congestion_visual}
\title{Compare Congestion Prices between two cities visually}
\usage{
compare_congestion_visual(
  df1 = read.csv(system.file("extdata", "San_Francisco_node_2023.csv", package =
    "CAISOWrapR")),
  df2 = read.csv(system.file("extdata", "Los_Angeles_node_2023.csv", package =
    "CAISOWrapR"))
)
}
\arguments{
\item{df1}{A dataframe of locational marginal prices for a year for one location. The default is San Francisco for 2023. Users can change this dataframe to analyze a different location or year by pulling from the \code{pulldata_node} function}

\item{df2}{A dataframe of locational marginal prices for a year for another location. The default is Los Angeles for 2023. Users can change this dataframe to analyze a different location or year by pulling from the \code{pulldata_node} function}
}
\value{
A ggplot object displaying the distribution of congestion prices by hour of the day for the two cities
}
\description{
This script reads in two datasets containing prices for two different cities over a year (or other specified time period) and compares the distribution of LMPs by hour of the day.
}
