% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_lmps.R
\name{compare_lmps}
\alias{compare_lmps}
\title{Compare LMPs}
\usage{
compare_lmps(df, loc1, loc2)
}
\arguments{
\item{df}{A dataframe of locational marginal prices with LMP names}

\item{loc1}{A string of the first location}

\item{loc2}{A string of the second location}
}
\value{
A named vector of the two locations and the difference in price
}
\description{
Function gets the input of two locations and a dataframe of locational marginal prices and returns the price comparison between the two locations.
}
\examples{
compare_lmps(df, "Boston", "New York")
}
