% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_lmps_visual.R
\name{compare_lmps_visual}
\alias{compare_lmps_visual}
\title{Compare LMPs between two cities visually}
\usage{
compare_lmps_visual(df1, df2)
}
\arguments{
\item{df1}{A tibble containing LMP data for the first city over a year}

\item{df2}{A tibble containing LMP data for the second city over a year}
}
\value{
A ggplot object displaying the distribution of LMPs by hour of the day for the two cities
}
\description{
This script reads in two datasets containing LMPs for two different cities over a year (or other specified time period) and compares the distribution of LMPs by hour of the day.
}
