% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caiso_api_wrapper.R
\name{convert_lmp_datatypes}
\alias{convert_lmp_datatypes}
\title{Convert LMP Data Types}
\usage{
convert_lmp_datatypes(data)
}
\arguments{
\item{data}{Input data frame}
}
\value{
Data frame with converted types
}
\description{
Convert API data types to appropriate R types
}
