% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caiso_api_wrapper.R
\name{execute_api_request}
\alias{execute_api_request}
\title{Execute API Request}
\usage{
execute_api_request(request)
}
\arguments{
\item{request}{httr2 request object}
}
\value{
Processed API response
}
\description{
Send API request and process response
}
