% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caiso_api_wrapper.R
\name{extract_first}
\alias{extract_first}
\title{Safely Extract First List Element}
\usage{
extract_first(x)
}
\arguments{
\item{x}{Input list or vector}
}
\value{
First element or NA
}
\description{
Safely extract first element from potentially nested lists
}
