% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caiso_api_wrapper.R
\name{get_lmp.CAISOClient}
\alias{get_lmp.CAISOClient}
\title{Fetch LMP data from the Grid Status API}
\usage{
get_lmp.CAISOClient(
  client,
  dataset = "caiso_lmp_real_time_5_min",
  start_tm = NULL,
  end_tm = NULL,
  columns = NULL,
  filter_column = NULL,
  filter_value = NULL,
  filter_operator = NULL,
  publish_tm = NULL,
  resample = NULL,
  resample_by = NULL,
  resample_function = "mean",
  limit = 5,
  page_size = NULL,
  tz = NULL,
  verbose = FALSE,
  use_cursor_pagination = TRUE
)
}
\arguments{
\item{client}{An CAISOClient object created with \code{\link{CAISOClient}}.}

\item{dataset}{The name of the dataset to fetch data from. Default is "caiso_lmp_real_time_5_min".
Options are}

\item{start_tm}{The start time for the data query. If not provided, the API will use the earliest available time.}

\item{end_tm}{The end time for the data query. If not provided, the API will use the latest available time.}

\item{columns}{A character vector of column names to include in the response. If not provided, all columns will be returned.}

\item{filter_column}{The name of the column to filter on. If provided, the filter_value and filter_operator parameters must also be provided.}

\item{filter_value}{The value to filter on. If provided, the filter_column and filter_operator parameters must also be provided.
If filter operator is "in", filter value should be a list of values to filter on.}

\item{filter_operator}{The operator to use for the filter. Options are "=", ">", "<", ">=", "<=", "in". If not provided, the default is "=".}

\item{publish_tm}{Controls the filtering on the dataset's publish time. The possible values are as follows:
\itemize{
\item "latest_report": Returns records only from the most recently published report
\item "latest": For any given timestamp, returns the most recently published record associated with it
\item timestamp str: Returns records that were published at the timestamp provided
\item None: No filtering
}}

\item{resample}{The frequency to resample the data to. Options are}

\item{resample_by}{A character vector of column names to group by when resampling the data.
By default resamples by the time index column. If resample is not provided, this parameter is ignored.}

\item{resample_function}{The function to use when resampling the data. Options are "mean", "sum", "min", "max", "stddev", "count", "variance".
Default is "mean".}

\item{limit}{The maximum number of records to return. Default is 5.}

\item{page_size}{The number of records to return per page. If not provided, the API will use the default page size allowed by subscription}

\item{tz}{The timezone to convert the timestamps to. Default is "UTC".}

\item{verbose}{If set to TRUE or "info", prints additional information. If set to "debug", prints more additional
debug information. If set to FALSE, does not print any additional information. Default is FALSE.}

\item{use_cursor_pagination}{A logical value indicating whether to use cursor-based pagination. Default is FALSE.}
}
\value{
A data frame containing the LMP data.
}
\description{
Fetch locational marginal price (LMP) data from the Grid Status API.
}
