% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_single_instance_prices.R
\name{gtable_add_padding}
\alias{gtable_add_padding}
\title{Add Padding to a gtable Object}
\usage{
gtable_add_padding(grob, padding)
}
\arguments{
\item{grob}{A \code{gtable} object to which padding will be added.}

\item{padding}{A unit object specifying the padding to apply (e.g., \code{unit(c(1, 1, 1, 1), "cm")}).}
}
\value{
A \code{gtable} object with the specified padding added.
}
\description{
This function adds padding to a \code{gtable} object by first inserting an invisible \code{rectGrob} around the table and then applying the specified padding.
It is a custom function to ensure that padding is applied correctly around the \code{grob} while preserving the layout of the original \code{gtable}.
}
\examples{
library(grid)
library(gtable)

# Create a simple table
example_table <- tableGrob(head(mtcars))

# Add padding to the table
padded_table <- gtable_add_padding(example_table, unit(c(1, 1, 1, 1), "cm"))

# Print the padded table
grid.newpage()
grid.draw(padded_table)

}
