% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isone_api_wrapper.R
\name{ISONEClient}
\alias{ISONEClient}
\title{ISONE API Wrapper Constructor}
\usage{
ISONEClient()
}
\value{
An ISONEClient object.
}
\description{
Construct an ISONEClient object for interacting with the ISO New England API.
}
\details{
In order to utilize ISO New England's RESTFul interface, you need basic authentication credentials
(username and password). To do so, set the environment variables \code{ISONE_API_USERNAME} and
\code{ISONE_API_PASSWORD} to your username and password for your ISO New England, respectively. If you do not have a
username and password, you can create a free account for ISO New England by visiting the following website:
\url{https://www.iso-ne.com/isoexpress/login?p_p_id=58&p_p_lifecycle=0&p_p_state=normal&p_p_mode=view&saveLastPath=0&_58_struts_action=\%2Flogin\%2Fcreate_account}.
}
