% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_demand.R
\name{merge_demand}
\alias{merge_demand}
\title{Merges a node with CAISO demand}
\usage{
merge_demand(node_data)
}
\arguments{
\item{node_data}{a data frame for a single node}
}
\value{
a data frame with the node data and the demand data
}
\description{
This function merges a node with the CAISO demand data by interval_start_tm and adds column demand for demand data
}
