% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_lmp_locations.R
\name{merge_lmp_locations}
\alias{merge_lmp_locations}
\title{Takes instance dataframe and merges with location and latitude longitude data}
\usage{
merge_lmp_locations(instance)
}
\arguments{
\item{instance}{dataframe of instance data}
}
\value{
dataframe with instance data and location data
}
\description{
Takes instance dataframe and merges with location and latitude longitude data
}
\examples{
merge_lmp_locations(instance)

}
