% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NE_map.R
\name{NE_map}
\alias{NE_map}
\title{create base map of NE}
\usage{
NE_map()
}
\value{
a map of the Northeastern states of the United States
}
\description{
function creates a map of the Northeastern states of the United States
}
\examples{
NE_map()

}
