% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caiso_api_wrapper.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess data from the Grid Status API}
\usage{
preprocess_data(data)
}
\arguments{
\item{data}{A data frame containing the raw data from the API.}
}
\value{
A data frame with the columns converted to the appropriate data types.
}
\description{
Preprocess the raw data from the Grid Status API by converting columns to the appropriate data types.
}
