% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_histogram.R
\name{price_histogram}
\alias{price_histogram}
\title{Price Histogram}
\usage{
price_histogram(df, bins)
}
\arguments{
\item{df}{A dataframe of locational marginal prices with LMP names}

\item{bins}{An integer of the number of bins to use in the histogram}
}
\value{
A histogram of the prices
}
\description{
This script generates a histogram of the price of the LMPs in the dataset.
}
\examples{
price_histogram(df, 30)
}
