% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_histogram.R
\name{price_violin}
\alias{price_violin}
\title{Price Violin}
\usage{
price_violin(df)
}
\arguments{
\item{df}{A dataframe of locational marginal prices with LMP names}
}
\value{
A violin plot of the prices with an overlaid boxplot
}
\description{
This script generates a violin plot of the LMP prices in the dataset,
with an overlaid boxplot showing the median, interquartile range, and whiskers.
}
\examples{
price_violin(df)
}
