% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caiso_api_wrapper.R
\name{process_lmp_response}
\alias{process_lmp_response}
\title{Process LMP Response}
\usage{
process_lmp_response(response)
}
\arguments{
\item{response}{API response object}
}
\value{
Processed data frame
}
\description{
Convert API response to a clean, type-converted data frame
}
