% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulldata_instance.R
\name{pulldata_instance}
\alias{pulldata_instance}
\title{Pull One Data Instnace}
\usage{
pulldata_instance(date, time, api_key)
}
\arguments{
\item{date}{Date in format "YYYY-MM-DD"}

\item{time}{Time in format "HH:MM"}

\item{api_key}{API key for the CAISO API}
}
\value{
A dataframe with the LMP and price breakdown for the given date and time
}
\description{
This function takes in a date and time (15 minute increment) and returns a dataframe of every LMP and the price breakdown
}
\examples{
pulldata_instance("2019-01-01", "00:00", api_key = api_key)
}
