% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulldata_node.R
\name{pulldata_node}
\alias{pulldata_node}
\title{Pull Data from One Node}
\usage{
pulldata_node(city = NULL, node = NULL, year = 2023, api_key)
}
\arguments{
\item{city}{city (automatically set to NULL)}

\item{node}{node (automatically set to NULL)}

\item{year}{year (2020-2024)}
}
\value{
a dataframe of the price every 15 minutes for that year in that location
}
\description{
This function takes in a city or a node and a year and returns a dataframe of the price every 15 minutes for that year in that location. Default is Palo Alto for 2024
}
\examples{
pulldata_node(city = "San Diego", year = 2020)
}
