% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caiso_api_wrapper.R
\name{safe_posixct}
\alias{safe_posixct}
\title{Safe POSIXct Conversion}
\usage{
safe_posixct(x)
}
\arguments{
\item{x}{Timestamp vector}
}
\value{
Converted POSIXct vector
}
\description{
Robust timestamp conversion
}
