% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_single_instance_prices.R,
%   R/summarize_lmps.R
\name{summarize_lmps}
\alias{summarize_lmps}
\title{Summarize Electricity Statistics}
\usage{
summarize_lmps(df)

summarize_lmps(df)
}
\arguments{
\item{df}{A dataframe of locational marginal prices}

\item{instance_df}{A dataframe of locational marginal prices}
}
\value{
A dataframe of summary statistics

A list of summary statistics
}
\description{
This function will give summary statistics for the current state of electricity prices. It will take in a dataframe of locational marginal prices (that has been pulled from an API) and return the average price around NE-ISO, the standard deviation in prices, the lowest price, the highest price, and the locations for those prices.

This function will give summary statistics for the current state of electricity prices. It will take in a dataframe of locational marginal prices (that has been pulled from an API) and return the average price around NE-ISO, the standard deviation in prices, the lowest price, the highest price, and the locations for those prices.
}
\examples{
# Example Data
example_df <- data.frame(
  lmp = c(20.5, 30.2, 25.1),
  location = c("Loc1", "Loc2", "Loc3")
)
summarize_lmps(example_df)
summarize_lmps(df)
}
