% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caiso_api_wrapper.R
\name{validate_client}
\alias{validate_client}
\title{Validate CAISO Client}
\usage{
validate_client(client)
}
\arguments{
\item{client}{CAISO client object}
}
\value{
Logical indicating client validity
}
\description{
Internal method to validate client object
}
