% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_node_data.R
\name{visualize_node}
\alias{visualize_node}
\title{Visualize Node Data}
\usage{
visualize_node(df)
}
\arguments{
\item{df}{a dataframe of a year's worth of 15-minute data for a node}
}
\value{
1 graph: facet grid of 4 line charts
}
\description{
This function takes in a dataframe of a year's worth of 15-minute data for a node
and visualizes the data in 4 graphs: average price per time (line chart),
price over time (line chart), average congestion per time (line chart), and
congestion over time (line chart).
}
