% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_single_node_prices.R
\name{visualize_node_demand_congestion}
\alias{visualize_node_demand_congestion}
\title{Visualize Node Data: Congestion}
\usage{
visualize_node_demand_congestion(
  node_df = read.csv("inst/extdata/San_Francisco_node_2023.csv")
)
}
\arguments{
\item{node_df}{A dataframe of locational marginal prices for a year for one node. The default for this is San Francisco 2023, but a user can input a different dataframe by running the \code{pulldata_node} function.}
}
\value{
A \code{ggplot2} object displaying the distribution of congestion by demand bucket (up to 40,000).
}
\description{
This function takes in a tibble of a year's worth of 15-minute data for a node and visualizes the data with a box and whisker plot of congestion grouped by load buckets.
}
