% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_single_node_prices.R,
%   R/visualize_node_data_seasonal.R
\name{visualize_node_seasonal}
\alias{visualize_node_seasonal}
\title{Visualize distribution of LMPs in one location by season}
\usage{
visualize_node_seasonal(df)

visualize_node_seasonal(df)
}
\arguments{
\item{df}{A tibble containing LMP data for the first city over a year}
}
\value{
A ggplot object displaying the distribution of LMPs by hour for the location over a day

A ggplot object displaying the distribution of LMPs by hour for the location over a day
}
\description{
This script reads in a dataframe of LMP data for a city over a year and visualizes the distribution of LMPs by hour for each season.

This script reads in a dataframe of LMP data for a city over a year and visualizes the distribution of LMPs by hour for each season.
}
