% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_single_node_prices.R
\name{visualize_node_seasonal_congestion}
\alias{visualize_node_seasonal_congestion}
\title{Visualize distribution of Congestion prices in one location by season}
\usage{
visualize_node_seasonal_congestion(
  node_df = read.csv(system.file("extdata", "San_Francisco_node_2023.csv", package =
    "CAISOWrapR"))
)
}
\arguments{
\item{node_df}{A dataframe of locational marginal prices for a year for one node. The default for this is San Francisco 2023, but a user can input a different dataframe by running the \code{pulldata_node} function.}
}
\value{
A \code{ggplot2} object displaying the distribution of LMPs by hour for the location over a day, split by season.
}
\description{
This script reads in a dataframe of LMP data for a city over a year and visualizes the distribution of LMPs by hour for each season.
}
