% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_single_instance_prices.R
\name{analyze_single_instance_data}
\alias{analyze_single_instance_data}
\title{Analyze Single Instance Data}
\usage{
analyze_single_instance_data(
  instance_df = read.csv(system.file("extdata", "instance_normal.csv", project =
    "CAISOWrapR")),
  output_pdf_path = "single_instance_data.pdf"
)
}
\arguments{
\item{instance_df}{A dataframe of locational marginal prices with LMP names. The default dataframe is a normal day in California at 1pm. Users can change this dataframe to analyze a different instance that pulled from the \code{pulldata_instance} function. Or, to visualize the highest peak load in the last three years, use: "inst/extdata/instance_peak_load.csv"}

\item{output_pdf_path}{The path to save the PDF file. The default path is "single_instance_data.pdf," but the user can modify this path to save the file in a different location.}
}
\value{
A PDF file with the summary statistics and violin plot.
}
\description{
This function calls the functions \code{summarize_lmps} and \code{price_violin} with an inputting dataframe of an instance of data (one date and 15 minute interval) and combines the results into a PDF for export. First, the user must call the \code{pulldata_instance} function to retrieve the data for the instance they want to analyze.
}
