% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isone_api_wrapper.R
\name{fetch_data}
\alias{fetch_data}
\title{Fetch data from the ISO New England API}
\usage{
fetch_data(client, endpoint, params = list())
}
\arguments{
\item{client}{An ISONEClient object created with \code{\link{ISONEClient}}.}

\item{endpoint}{The API endpoint to fetch data from.}

\item{params}{A list of query parameters to include in the request.}
}
\value{
A list representing the JSON response from the API.
}
\description{
Fetch data from the ISO New England API using the provided client, endpoint, and query parameters.
}
