% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_single_instance_prices.R
\name{price_violin}
\alias{price_violin}
\title{Price Violin}
\usage{
price_violin(
  instance_df = read.csv(system.file("extdata", "instance_normal.csv", project =
    "CAISOWrapR"))
)
}
\arguments{
\item{instance_df}{A dataframe of locational marginal prices with LMP names. The default dataframe is a normal day in California at 1pm. Users can change this dataframe to analyze a different instance that pulled from the \code{pulldata_instance} function. Or, to visualize the highest peak load in the last three years, use: "inst/extdata/instance_peak_load.csv"}
}
\value{
A violin plot of the prices with an overlaid boxplot.
}
\description{
This function generates a violin plot of the LMP prices in the dataset,
with an overlaid boxplot showing the median, interquartile range, and whiskers.
}
\examples{
# Example Data
example_df <- data.frame(
  lmp = c(20.5, 30.2, 25.1),
  location = c("Loc1", "Loc2", "Loc3")
)
price_violin(example_df)
}
