% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_lmps.R
\name{summarize_lmps}
\alias{summarize_lmps}
\title{Summarize Electricity Statistics}
\usage{
summarize_lmps(df)
}
\arguments{
\item{df}{A dataframe of locational marginal prices}
}
\value{
A list of summary statistics
}
\description{
This function will give summary statistics for the current state of electricity prices. It will take in a dataframe of locational marginal prices (that has been pulled from an API) and return the average price around NE-ISO, the standard deviation in prices, the lowest price, the highest price, and the locations for those prices.
}
\examples{
summarize_lmps(df)
}
