% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_single_instance_prices.R
\name{summarize_lmps}
\alias{summarize_lmps}
\title{Summarize Electricity Statistics}
\usage{
summarize_lmps(instance_df = read.csv("inst/extdata/instance_normal.csv"))
}
\arguments{
\item{instance_df}{A dataframe of locational marginal prices.}
}
\value{
A dataframe of summary statistics. The default dataframe is a normal day in California at 1pm. Users can change this dataframe to analyze a different instance that pulled from the \code{pulldata_instance} function. Or, to visualize the highest peak load in the last three years, use: "inst/extdata/instance_peak_load.csv"
}
\description{
This function provides summary statistics for the current state of electricity prices. It takes in a dataframe of locational marginal prices (LMPs) and returns the average price, the standard deviation of prices, the lowest price, the highest price, and the locations for those prices.
}
\examples{
# Example Data
example_df <- data.frame(
  lmp = c(20.5, 30.2, 25.1),
  location = c("Loc1", "Loc2", "Loc3")
)
summarize_lmps(example_df)
}
