% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_single_node_prices.R,
%   R/visualize_node_data.R
\name{visualize_node}
\alias{visualize_node}
\title{Visualize Node Data}
\usage{
visualize_node(df)

visualize_node(df)
}
\arguments{
\item{df}{a dataframe of a year's worth of 15-minute data for a node}

\item{node_df}{a dataframe of a year's worth of 15-minute data for a node}
}
\value{
1 graph

1 graph
}
\description{
This function takes in a dataframe of a year's worth of 15-minute data for a node
and visualizes the data in a box and whisker plot by interval to see how the price distribution ranges
over a day

This function takes in a dataframe of a year's worth of 15-minute data for a node
and visualizes the data in a box and whisker plot by interval to see how the price distribution ranges
over a day
}
